@@ -6,34 +6,34 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  tflite: ^1.1.1
-  percent_indicator: ^2.1.1+1
-  image_picker: ^0.7.2+1
+  tflite: ^1.1.2
+  percent_indicator: ^3.0.1
+  image_picker: ^0.8.3+3
   moor_flutter: ^4.0.0
-  flutter_mobx: ^1.1.0+2
-  google_maps_flutter: ^2.0.1
-  geolocator: ^7.0.1
-  flutter_polyline_points: ^0.2.6
-  path_provider: ^2.0.1
-  sqlite3_flutter_libs: ^0.2.0
+  flutter_mobx: ^2.0.2
+  google_maps_flutter: ^2.0.6
+  geolocator: ^7.5.0
+  flutter_polyline_points: ^1.0.0
+  path_provider: ^2.0.2
+  sqlite3_flutter_libs: ^0.5.0
   intl: ^0.17.0
-  flutter_slidable: ^0.5.7
-  material_design_icons_flutter: ^4.0.5955
+  flutter_slidable: ^0.6.0
+  material_design_icons_flutter: ^5.0.5955-rc.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   moor_generator:
-  mobx_codegen: ^1.1.0
-  build_runner: ^1.11.0
+  mobx_codegen:
+  build_runner:
 
 dependencies_overrides:
   # mobx_codegen: ^2.0.0-nullsafety.2
