@@ -8,106 +8,87 @@ part of 'database.dart';
 
 // ignore_for_file: unnecessary_brace_in_string_interps, unnecessary_this
 class Analyze extends DataClass implements Insertable<Analyze> {
-  final int id;
+  final int? id;
   final String imageDir;
-  final String result;
+  final String? result;
   final String species;
   final String description;
   final DateTime date;
-  final String note;
-  final double latitude;
-  final double longitude;
+  final String? note;
+  final double? latitude;
+  final double? longitude;
   final double percentage;
   Analyze(
       {this.id,
-      @required this.imageDir,
+      required this.imageDir,
       this.result,
-      @required this.species,
-      @required this.description,
-      @required this.date,
+      required this.species,
+      required this.description,
+      required this.date,
       this.note,
       this.latitude,
       this.longitude,
-      @required this.percentage});
+      required this.percentage});
   factory Analyze.fromData(Map<String, dynamic> data, GeneratedDatabase db,
-      {String prefix}) {
+      {String? prefix}) {
     final effectivePrefix = prefix ?? '';
-    final intType = db.typeSystem.forDartType<int>();
-    final stringType = db.typeSystem.forDartType<String>();
-    final dateTimeType = db.typeSystem.forDartType<DateTime>();
-    final doubleType = db.typeSystem.forDartType<double>();
     return Analyze(
-      id: intType.mapFromDatabaseResponse(data['${effectivePrefix}id']),
-      imageDir: stringType
-          .mapFromDatabaseResponse(data['${effectivePrefix}image_dir']),
-      result:
-          stringType.mapFromDatabaseResponse(data['${effectivePrefix}result']),
-      species:
-          stringType.mapFromDatabaseResponse(data['${effectivePrefix}species']),
-      description: stringType
-          .mapFromDatabaseResponse(data['${effectivePrefix}description']),
-      date:
-          dateTimeType.mapFromDatabaseResponse(data['${effectivePrefix}date']),
-      note: stringType.mapFromDatabaseResponse(data['${effectivePrefix}note']),
-      latitude: doubleType
+      id: const IntType().mapFromDatabaseResponse(data['${effectivePrefix}id']),
+      imageDir: const StringType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}image_dir'])!,
+      result: const StringType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}result']),
+      species: const StringType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}species'])!,
+      description: const StringType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}description'])!,
+      date: const DateTimeType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}date'])!,
+      note: const StringType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}note']),
+      latitude: const RealType()
           .mapFromDatabaseResponse(data['${effectivePrefix}latitude']),
-      longitude: doubleType
+      longitude: const RealType()
           .mapFromDatabaseResponse(data['${effectivePrefix}longitude']),
-      percentage: doubleType
-          .mapFromDatabaseResponse(data['${effectivePrefix}percentage']),
+      percentage: const RealType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}percentage'])!,
     );
   }
   @override
   Map<String, Expression> toColumns(bool nullToAbsent) {
     final map = <String, Expression>{};
     if (!nullToAbsent || id != null) {
-      map['id'] = Variable<int>(id);
+      map['id'] = Variable<int?>(id);
     }
-    if (!nullToAbsent || imageDir != null) {
-      map['image_dir'] = Variable<String>(imageDir);
-    }
+    map['image_dir'] = Variable<String>(imageDir);
     if (!nullToAbsent || result != null) {
-      map['result'] = Variable<String>(result);
+      map['result'] = Variable<String?>(result);
     }
-    if (!nullToAbsent || species != null) {
-      map['species'] = Variable<String>(species);
-    }
-    if (!nullToAbsent || description != null) {
-      map['description'] = Variable<String>(description);
-    }
-    if (!nullToAbsent || date != null) {
-      map['date'] = Variable<DateTime>(date);
-    }
+    map['species'] = Variable<String>(species);
+    map['description'] = Variable<String>(description);
+    map['date'] = Variable<DateTime>(date);
     if (!nullToAbsent || note != null) {
-      map['note'] = Variable<String>(note);
+      map['note'] = Variable<String?>(note);
     }
     if (!nullToAbsent || latitude != null) {
-      map['latitude'] = Variable<double>(latitude);
+      map['latitude'] = Variable<double?>(latitude);
     }
     if (!nullToAbsent || longitude != null) {
-      map['longitude'] = Variable<double>(longitude);
+      map['longitude'] = Variable<double?>(longitude);
     }
-    if (!nullToAbsent || percentage != null) {
-      map['percentage'] = Variable<double>(percentage);
-    }
+    map['percentage'] = Variable<double>(percentage);
     return map;
   }
 
   AnalysisCompanion toCompanion(bool nullToAbsent) {
     return AnalysisCompanion(
       id: id == null && nullToAbsent ? const Value.absent() : Value(id),
-      imageDir: imageDir == null && nullToAbsent
-          ? const Value.absent()
-          : Value(imageDir),
+      imageDir: Value(imageDir),
       result:
           result == null && nullToAbsent ? const Value.absent() : Value(result),
-      species: species == null && nullToAbsent
-          ? const Value.absent()
-          : Value(species),
-      description: description == null && nullToAbsent
-          ? const Value.absent()
-          : Value(description),
-      date: date == null && nullToAbsent ? const Value.absent() : Value(date),
+      species: Value(species),
+      description: Value(description),
+      date: Value(date),
       note: note == null && nullToAbsent ? const Value.absent() : Value(note),
       latitude: latitude == null && nullToAbsent
           ? const Value.absent()
@@ -115,56 +96,54 @@ class Analyze extends DataClass implements Insertable<Analyze> {
       longitude: longitude == null && nullToAbsent
           ? const Value.absent()
           : Value(longitude),
-      percentage: percentage == null && nullToAbsent
-          ? const Value.absent()
-          : Value(percentage),
+      percentage: Value(percentage),
     );
   }
 
   factory Analyze.fromJson(Map<String, dynamic> json,
-      {ValueSerializer serializer}) {
+      {ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return Analyze(
-      id: serializer.fromJson<int>(json['id']),
+      id: serializer.fromJson<int?>(json['id']),
       imageDir: serializer.fromJson<String>(json['imageDir']),
-      result: serializer.fromJson<String>(json['result']),
+      result: serializer.fromJson<String?>(json['result']),
       species: serializer.fromJson<String>(json['species']),
       description: serializer.fromJson<String>(json['description']),
       date: serializer.fromJson<DateTime>(json['date']),
-      note: serializer.fromJson<String>(json['note']),
-      latitude: serializer.fromJson<double>(json['latitude']),
-      longitude: serializer.fromJson<double>(json['longitude']),
+      note: serializer.fromJson<String?>(json['note']),
+      latitude: serializer.fromJson<double?>(json['latitude']),
+      longitude: serializer.fromJson<double?>(json['longitude']),
       percentage: serializer.fromJson<double>(json['percentage']),
     );
   }
   @override
-  Map<String, dynamic> toJson({ValueSerializer serializer}) {
+  Map<String, dynamic> toJson({ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return <String, dynamic>{
-      'id': serializer.toJson<int>(id),
+      'id': serializer.toJson<int?>(id),
       'imageDir': serializer.toJson<String>(imageDir),
-      'result': serializer.toJson<String>(result),
+      'result': serializer.toJson<String?>(result),
       'species': serializer.toJson<String>(species),
       'description': serializer.toJson<String>(description),
       'date': serializer.toJson<DateTime>(date),
-      'note': serializer.toJson<String>(note),
-      'latitude': serializer.toJson<double>(latitude),
-      'longitude': serializer.toJson<double>(longitude),
+      'note': serializer.toJson<String?>(note),
+      'latitude': serializer.toJson<double?>(latitude),
+      'longitude': serializer.toJson<double?>(longitude),
       'percentage': serializer.toJson<double>(percentage),
     };
   }
 
   Analyze copyWith(
-          {int id,
-          String imageDir,
-          String result,
-          String species,
-          String description,
-          DateTime date,
-          String note,
-          double latitude,
-          double longitude,
-          double percentage}) =>
+          {int? id,
+          String? imageDir,
+          String? result,
+          String? species,
+          String? description,
+          DateTime? date,
+          String? note,
+          double? latitude,
+          double? longitude,
+          double? percentage}) =>
       Analyze(
         id: id ?? this.id,
         imageDir: imageDir ?? this.imageDir,
@@ -214,7 +193,7 @@ class Analyze extends DataClass implements Insertable<Analyze> {
                                   $mrjc(longitude.hashCode,
                                       percentage.hashCode))))))))));
   @override
-  bool operator ==(dynamic other) =>
+  bool operator ==(Object other) =>
       identical(this, other) ||
       (other is Analyze &&
           other.id == this.id &&
@@ -230,15 +209,15 @@ class Analyze extends DataClass implements Insertable<Analyze> {
 }
 
 class AnalysisCompanion extends UpdateCompanion<Analyze> {
-  final Value<int> id;
+  final Value<int?> id;
   final Value<String> imageDir;
-  final Value<String> result;
+  final Value<String?> result;
   final Value<String> species;
   final Value<String> description;
   final Value<DateTime> date;
-  final Value<String> note;
-  final Value<double> latitude;
-  final Value<double> longitude;
+  final Value<String?> note;
+  final Value<double?> latitude;
+  final Value<double?> longitude;
   final Value<double> percentage;
   const AnalysisCompanion({
     this.id = const Value.absent(),
@@ -254,31 +233,31 @@ class AnalysisCompanion extends UpdateCompanion<Analyze> {
   });
   AnalysisCompanion.insert({
     this.id = const Value.absent(),
-    @required String imageDir,
+    required String imageDir,
     this.result = const Value.absent(),
-    @required String species,
-    @required String description,
-    @required DateTime date,
+    required String species,
+    required String description,
+    required DateTime date,
     this.note = const Value.absent(),
     this.latitude = const Value.absent(),
     this.longitude = const Value.absent(),
-    @required double percentage,
+    required double percentage,
   })  : imageDir = Value(imageDir),
         species = Value(species),
         description = Value(description),
         date = Value(date),
         percentage = Value(percentage);
   static Insertable<Analyze> custom({
-    Expression<int> id,
-    Expression<String> imageDir,
-    Expression<String> result,
-    Expression<String> species,
-    Expression<String> description,
-    Expression<DateTime> date,
-    Expression<String> note,
-    Expression<double> latitude,
-    Expression<double> longitude,
-    Expression<double> percentage,
+    Expression<int?>? id,
+    Expression<String>? imageDir,
+    Expression<String?>? result,
+    Expression<String>? species,
+    Expression<String>? description,
+    Expression<DateTime>? date,
+    Expression<String?>? note,
+    Expression<double?>? latitude,
+    Expression<double?>? longitude,
+    Expression<double>? percentage,
   }) {
     return RawValuesInsertable({
       if (id != null) 'id': id,
@@ -295,16 +274,16 @@ class AnalysisCompanion extends UpdateCompanion<Analyze> {
   }
 
   AnalysisCompanion copyWith(
-      {Value<int> id,
-      Value<String> imageDir,
-      Value<String> result,
-      Value<String> species,
-      Value<String> description,
-      Value<DateTime> date,
-      Value<String> note,
-      Value<double> latitude,
-      Value<double> longitude,
-      Value<double> percentage}) {
+      {Value<int?>? id,
+      Value<String>? imageDir,
+      Value<String?>? result,
+      Value<String>? species,
+      Value<String>? description,
+      Value<DateTime>? date,
+      Value<String?>? note,
+      Value<double?>? latitude,
+      Value<double?>? longitude,
+      Value<double>? percentage}) {
     return AnalysisCompanion(
       id: id ?? this.id,
       imageDir: imageDir ?? this.imageDir,
@@ -323,13 +302,13 @@ class AnalysisCompanion extends UpdateCompanion<Analyze> {
   Map<String, Expression> toColumns(bool nullToAbsent) {
     final map = <String, Expression>{};
     if (id.present) {
-      map['id'] = Variable<int>(id.value);
+      map['id'] = Variable<int?>(id.value);
     }
     if (imageDir.present) {
       map['image_dir'] = Variable<String>(imageDir.value);
     }
     if (result.present) {
-      map['result'] = Variable<String>(result.value);
+      map['result'] = Variable<String?>(result.value);
     }
     if (species.present) {
       map['species'] = Variable<String>(species.value);
@@ -341,13 +320,13 @@ class AnalysisCompanion extends UpdateCompanion<Analyze> {
       map['date'] = Variable<DateTime>(date.value);
     }
     if (note.present) {
-      map['note'] = Variable<String>(note.value);
+      map['note'] = Variable<String?>(note.value);
     }
     if (latitude.present) {
-      map['latitude'] = Variable<double>(latitude.value);
+      map['latitude'] = Variable<double?>(latitude.value);
     }
     if (longitude.present) {
-      map['longitude'] = Variable<double>(longitude.value);
+      map['longitude'] = Variable<double?>(longitude.value);
     }
     if (percentage.present) {
       map['percentage'] = Variable<double>(percentage.value);
@@ -375,127 +354,51 @@ class AnalysisCompanion extends UpdateCompanion<Analyze> {
 
 class $AnalysisTable extends Analysis with TableInfo<$AnalysisTable, Analyze> {
   final GeneratedDatabase _db;
-  final String _alias;
+  final String? _alias;
   $AnalysisTable(this._db, [this._alias]);
   final VerificationMeta _idMeta = const VerificationMeta('id');
-  GeneratedIntColumn _id;
-  @override
-  GeneratedIntColumn get id => _id ??= _constructId();
-  GeneratedIntColumn _constructId() {
-    return GeneratedIntColumn('id', $tableName, true,
-        hasAutoIncrement: true, declaredAsPrimaryKey: true);
-  }
-
+  late final GeneratedColumn<int?> id = GeneratedColumn<int?>(
+      'id', aliasedName, true,
+      typeName: 'INTEGER',
+      requiredDuringInsert: false,
+      defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
   final VerificationMeta _imageDirMeta = const VerificationMeta('imageDir');
-  GeneratedTextColumn _imageDir;
-  @override
-  GeneratedTextColumn get imageDir => _imageDir ??= _constructImageDir();
-  GeneratedTextColumn _constructImageDir() {
-    return GeneratedTextColumn(
-      'image_dir',
-      $tableName,
-      false,
-    );
-  }
-
+  late final GeneratedColumn<String?> imageDir = GeneratedColumn<String?>(
+      'image_dir', aliasedName, false,
+      typeName: 'TEXT', requiredDuringInsert: true);
   final VerificationMeta _resultMeta = const VerificationMeta('result');
-  GeneratedTextColumn _result;
-  @override
-  GeneratedTextColumn get result => _result ??= _constructResult();
-  GeneratedTextColumn _constructResult() {
-    return GeneratedTextColumn(
-      'result',
-      $tableName,
-      true,
-    );
-  }
-
+  late final GeneratedColumn<String?> result = GeneratedColumn<String?>(
+      'result', aliasedName, true,
+      typeName: 'TEXT', requiredDuringInsert: false);
   final VerificationMeta _speciesMeta = const VerificationMeta('species');
-  GeneratedTextColumn _species;
-  @override
-  GeneratedTextColumn get species => _species ??= _constructSpecies();
-  GeneratedTextColumn _constructSpecies() {
-    return GeneratedTextColumn(
-      'species',
-      $tableName,
-      false,
-    );
-  }
-
+  late final GeneratedColumn<String?> species = GeneratedColumn<String?>(
+      'species', aliasedName, false,
+      typeName: 'TEXT', requiredDuringInsert: true);
   final VerificationMeta _descriptionMeta =
       const VerificationMeta('description');
-  GeneratedTextColumn _description;
-  @override
-  GeneratedTextColumn get description =>
-      _description ??= _constructDescription();
-  GeneratedTextColumn _constructDescription() {
-    return GeneratedTextColumn(
-      'description',
-      $tableName,
-      false,
-    );
-  }
-
+  late final GeneratedColumn<String?> description = GeneratedColumn<String?>(
+      'description', aliasedName, false,
+      typeName: 'TEXT', requiredDuringInsert: true);
   final VerificationMeta _dateMeta = const VerificationMeta('date');
-  GeneratedDateTimeColumn _date;
-  @override
-  GeneratedDateTimeColumn get date => _date ??= _constructDate();
-  GeneratedDateTimeColumn _constructDate() {
-    return GeneratedDateTimeColumn(
-      'date',
-      $tableName,
-      false,
-    );
-  }
-
+  late final GeneratedColumn<DateTime?> date = GeneratedColumn<DateTime?>(
+      'date', aliasedName, false,
+      typeName: 'INTEGER', requiredDuringInsert: true);
   final VerificationMeta _noteMeta = const VerificationMeta('note');
-  GeneratedTextColumn _note;
-  @override
-  GeneratedTextColumn get note => _note ??= _constructNote();
-  GeneratedTextColumn _constructNote() {
-    return GeneratedTextColumn(
-      'note',
-      $tableName,
-      true,
-    );
-  }
-
+  late final GeneratedColumn<String?> note = GeneratedColumn<String?>(
+      'note', aliasedName, true,
+      typeName: 'TEXT', requiredDuringInsert: false);
   final VerificationMeta _latitudeMeta = const VerificationMeta('latitude');
-  GeneratedRealColumn _latitude;
-  @override
-  GeneratedRealColumn get latitude => _latitude ??= _constructLatitude();
-  GeneratedRealColumn _constructLatitude() {
-    return GeneratedRealColumn(
-      'latitude',
-      $tableName,
-      true,
-    );
-  }
-
+  late final GeneratedColumn<double?> latitude = GeneratedColumn<double?>(
+      'latitude', aliasedName, true,
+      typeName: 'REAL', requiredDuringInsert: false);
   final VerificationMeta _longitudeMeta = const VerificationMeta('longitude');
-  GeneratedRealColumn _longitude;
-  @override
-  GeneratedRealColumn get longitude => _longitude ??= _constructLongitude();
-  GeneratedRealColumn _constructLongitude() {
-    return GeneratedRealColumn(
-      'longitude',
-      $tableName,
-      true,
-    );
-  }
-
+  late final GeneratedColumn<double?> longitude = GeneratedColumn<double?>(
+      'longitude', aliasedName, true,
+      typeName: 'REAL', requiredDuringInsert: false);
   final VerificationMeta _percentageMeta = const VerificationMeta('percentage');
-  GeneratedRealColumn _percentage;
-  @override
-  GeneratedRealColumn get percentage => _percentage ??= _constructPercentage();
-  GeneratedRealColumn _constructPercentage() {
-    return GeneratedRealColumn(
-      'percentage',
-      $tableName,
-      false,
-    );
-  }
-
+  late final GeneratedColumn<double?> percentage = GeneratedColumn<double?>(
+      'percentage', aliasedName, false,
+      typeName: 'REAL', requiredDuringInsert: true);
   @override
   List<GeneratedColumn> get $columns => [
         id,
@@ -510,32 +413,30 @@ class $AnalysisTable extends Analysis with TableInfo<$AnalysisTable, Analyze> {
         percentage
       ];
   @override
-  $AnalysisTable get asDslTable => this;
+  String get aliasedName => _alias ?? 'analysis';
   @override
-  String get $tableName => _alias ?? 'analysis';
-  @override
-  final String actualTableName = 'analysis';
+  String get actualTableName => 'analysis';
   @override
   VerificationContext validateIntegrity(Insertable<Analyze> instance,
       {bool isInserting = false}) {
     final context = VerificationContext();
     final data = instance.toColumns(true);
     if (data.containsKey('id')) {
-      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id'], _idMeta));
+      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id']!, _idMeta));
     }
     if (data.containsKey('image_dir')) {
       context.handle(_imageDirMeta,
-          imageDir.isAcceptableOrUnknown(data['image_dir'], _imageDirMeta));
+          imageDir.isAcceptableOrUnknown(data['image_dir']!, _imageDirMeta));
     } else if (isInserting) {
       context.missing(_imageDirMeta);
     }
     if (data.containsKey('result')) {
       context.handle(_resultMeta,
-          result.isAcceptableOrUnknown(data['result'], _resultMeta));
+          result.isAcceptableOrUnknown(data['result']!, _resultMeta));
     }
     if (data.containsKey('species')) {
       context.handle(_speciesMeta,
-          species.isAcceptableOrUnknown(data['species'], _speciesMeta));
+          species.isAcceptableOrUnknown(data['species']!, _speciesMeta));
     } else if (isInserting) {
       context.missing(_speciesMeta);
     }
@@ -543,33 +444,33 @@ class $AnalysisTable extends Analysis with TableInfo<$AnalysisTable, Analyze> {
       context.handle(
           _descriptionMeta,
           description.isAcceptableOrUnknown(
-              data['description'], _descriptionMeta));
+              data['description']!, _descriptionMeta));
     } else if (isInserting) {
       context.missing(_descriptionMeta);
     }
     if (data.containsKey('date')) {
       context.handle(
-          _dateMeta, date.isAcceptableOrUnknown(data['date'], _dateMeta));
+          _dateMeta, date.isAcceptableOrUnknown(data['date']!, _dateMeta));
     } else if (isInserting) {
       context.missing(_dateMeta);
     }
     if (data.containsKey('note')) {
       context.handle(
-          _noteMeta, note.isAcceptableOrUnknown(data['note'], _noteMeta));
+          _noteMeta, note.isAcceptableOrUnknown(data['note']!, _noteMeta));
     }
     if (data.containsKey('latitude')) {
       context.handle(_latitudeMeta,
-          latitude.isAcceptableOrUnknown(data['latitude'], _latitudeMeta));
+          latitude.isAcceptableOrUnknown(data['latitude']!, _latitudeMeta));
     }
     if (data.containsKey('longitude')) {
       context.handle(_longitudeMeta,
-          longitude.isAcceptableOrUnknown(data['longitude'], _longitudeMeta));
+          longitude.isAcceptableOrUnknown(data['longitude']!, _longitudeMeta));
     }
     if (data.containsKey('percentage')) {
       context.handle(
           _percentageMeta,
           percentage.isAcceptableOrUnknown(
-              data['percentage'], _percentageMeta));
+              data['percentage']!, _percentageMeta));
     } else if (isInserting) {
       context.missing(_percentageMeta);
     }
@@ -579,9 +480,9 @@ class $AnalysisTable extends Analysis with TableInfo<$AnalysisTable, Analyze> {
   @override
   Set<GeneratedColumn> get $primaryKey => {id};
   @override
-  Analyze map(Map<String, dynamic> data, {String tablePrefix}) {
-    final effectivePrefix = tablePrefix != null ? '$tablePrefix.' : null;
-    return Analyze.fromData(data, _db, prefix: effectivePrefix);
+  Analyze map(Map<String, dynamic> data, {String? tablePrefix}) {
+    return Analyze.fromData(data, _db,
+        prefix: tablePrefix != null ? '$tablePrefix.' : null);
   }
 
   @override
@@ -592,8 +493,7 @@ class $AnalysisTable extends Analysis with TableInfo<$AnalysisTable, Analyze> {
 
 abstract class _$MyDatabase extends GeneratedDatabase {
   _$MyDatabase(QueryExecutor e) : super(SqlTypeSystem.defaultInstance, e);
-  $AnalysisTable _analysis;
-  $AnalysisTable get analysis => _analysis ??= $AnalysisTable(this);
+  late final $AnalysisTable analysis = $AnalysisTable(this);
   @override
   Iterable<TableInfo> get allTables => allSchemaEntities.whereType<TableInfo>();
   @override
